poboljšan kalkulator
sada se mogu upisati i decimalne vrijednosti, korisnik može preciznije izračunati svoj indeks tjelesne mase. Ne mogu se više slati prazne vrijednosti, riješeno je  na način da se u html kodu upisalo da je potrebno upisati te vrijednosti a ne da ostanu prazne.

+++ PhysicalActivityMVC/src/main/java/com/josip/physical/activity/indeks_tjelesne_mase/BMIReprository.java
@@ -1,5 +1,5 @@
 package com.josip.physical.activity.indeks_tjelesne_mase;
 
 public interface BMIReprository {
-double izracunajBMI(int masa,double visina_osobe);
+double izracunajBMI(double masa,double visina_osobe);
 }
